main.py
import asyncio
import logging
from pathlib import Path

import uvicorn
from dotenv import load_dotenv
from fastapi import FastAPI

# 配置日志
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

# 加载环境变量
env_path = Path('.') / '.env'
load_dotenv(dotenv_path=env_path)

# 创建FastAPI应用
app = FastAPI(
    title="AI Crypto Trading Bot",
    description="基于大模型驱动的智能加密货币交易系统",
    version="1.0.0"
)

@app.on_event("startup")
async def startup_event():
    """
    应用启动时的初始化操作
    """
    logger.info("正在初始化交易系统...")
    # TODO: 初始化数据库连接
    # TODO: 初始化交易所API客户端
    # TODO: 启动数据采集服务
    # TODO: 启动AI分析服务
    # TODO: 启动交易执行服务

@app.on_event("shutdown")
async def shutdown_event():
    """
    应用关闭时的清理操作
    """
    logger.info("正在关闭交易系统...")
    # TODO: 关闭所有连接和服务

@app.get("/")
async def root():
    """
    API根路径
    """
    return {
        "status": "running",
        "service": "AI Crypto Trading Bot",
        "version": "1.0.0"
    }

def main():
    """
    主程序入口
    """
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=8000,
        reload=True  # 开发模式下启用热重载
    )

if __name__ == "__main__":
    main() 