requirements.txt
# 基础依赖
python-binance==1.0.19  # 币安API客户端
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0

# AI模型相关
openai==1.3.7
transformers==4.35.2
torch==2.1.1

# 数据库
pymongo==4.6.1
redis==5.0.1

# Web框架和API
fastapi==0.104.1
uvicorn==0.24.0.post1

# 数据可视化
plotly==5.18.0
dash==2.14.1

# 监控和日志
prometheus-client==0.19.0
python-logging-loki==0.3.1

# 测试工具
pytest==7.4.3
pytest-asyncio==0.21.1

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 